iter1.21 | Перенесена логика обработки начислений в новый пакет.
Вся логика работы с системой начислений перемещена из пула в отдельный пакет, что улучшает разделение ответственности и поддерживаемость кода. Это включает в себя ошибки, паузы и запросы к внешнему сервису.

Измененные файлы:
- internal/ordercheck/pool.go
- internal/accrual/proxy.go
